requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
ultralytics==8.0.237
scikit-learn==1.4.2
imutils==0.5.4
